# src/UI/ApiClient/LANdalf-OAS-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LANdalf.API | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7206/"
    }
  ],
  "paths": {
    "/api/v1/pc-devices": {
      "get": {
        "tags": [
          "LANdalf-api"
        ],
        "summary": "Get all PcDevices.",
        "description": "Returns an array of PcDevices.",
        "operationId": "GetAllPcDevices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PcDeviceDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pc-devices/{id}": {
      "get": {
        "tags": [
          "LANdalf-api"
        ],
        "summary": "Get a specific PcDevices.",
        "description": "Returns a PcDevice.",
        "operationId": "GetPcDeviceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PcDeviceDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pc-devices/add": {
      "post": {
        "tags": [
          "LANdalf-api"
        ],
        "summary": "Add a new PcDevice.",
        "description": "Returns a Created-Result.",
        "operationId": "AddPcDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PcCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pc-devices/{id}/set": {
      "post": {
        "tags": [
          "LANdalf-api"
        ],
        "summary": "Updates an existing PcDevice.",
        "description": "Returns a NoContent-Result.",
        "operationId": "SetPcDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PcDeviceDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pc-devices/{id}/delete": {
      "post": {
        "tags": [
          "LANdalf-api"
        ],
        "summary": "Deletes an existing PcDevice.",
        "description": "Returns a NoContent-Result.",
        "operationId": "DeletePcDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pc-devices/{id}/wake": {
      "post": {
        "tags": [
          "LANdalf-api"
        ],
        "summary": "Wakes an PcDevice.",
        "description": "Returns a Ok-Result.",
        "operationId": "WakePcDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PcCreateDto": {
        "required": [
          "name",
          "macAddress",
          "ipAddress",
          "broadcastAddress"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "macAddress": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string"
          },
          "broadcastAddress": {
            "type": "string"
          }
        }
      },
      "PcDeviceDTO": {
        "required": [
          "id",
          "name",
          "macAddress",
          "ipAddress",
          "broadcastAddress",
          "isOnline"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "macAddress": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string"
          },
          "broadcastAddress": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "LANdalf-api"
    }
  ]
}